API

API

AverageShiftedHistograms.Kernels

AverageShiftedHistograms.ash

AverageShiftedHistograms.ash

AverageShiftedHistograms.ash!

AverageShiftedHistograms.extendrange

AverageShiftedHistograms.histdensity

AverageShiftedHistograms.nout

AverageShiftedHistograms.xy

Base.quantile

StatsBase.nobs

AverageShiftedHistograms.ash! — Method.

ash!(o::Ash; kw...)
ash!(o::Ash, newdata; kw...)

Update an Ash estimate with new data, smoothing parameter (keyword m), or kernel (keyword kernel):

source

AverageShiftedHistograms.ash — Method.

ash(x; rngx::Range = extendrange(y), m = 5, kernel = Kernels.biweight)

Fit an average shifted histogram where:

y is the data

x is a range of values where the density should be estimated

m is a smoothing parameter. It is the number of adjacent histogram bins on either side used to estimate the density.

kernel is the kernel used to smooth the estimate

Make changes to the estimate (add more data, change kernel, or change smoothing parameter):

ash!(o::Ash; kernel = newkernel, m = newm)
ash!(o::Ash, y; kernel = newkernel, m = newm)

source

AverageShiftedHistograms.extendrange — Function.

extendrange(x, s = .5, n = 200)

Create a LinSpace of length n starting at s standard deviations below minimum(x) and ending at s standard deviations above maximum(x)

source

AverageShiftedHistograms.nout — Method.

return the number of observations that fell outside of the histogram range

source

AverageShiftedHistograms.xy — Method.

return the range and density as a tuple

source

StatsBase.nobs — Method.

return the number of observations

source

AverageShiftedHistograms.histdensity — Method.

return the histogram values as a density (intergrates to 1)

source

Base.quantile — Method.

quantile(o::Ash, q::Real)

Return the approximate q-th quantile from the Ash density.

source