API

AverageShiftedHistograms.ash

AverageShiftedHistograms.ash

AverageShiftedHistograms.ash!

AverageShiftedHistograms.cdf

AverageShiftedHistograms.extendrange

AverageShiftedHistograms.histdensity

AverageShiftedHistograms.nout

AverageShiftedHistograms.pdf

AverageShiftedHistograms.pdf

AverageShiftedHistograms.xy

AverageShiftedHistograms.xyz

StatsAPI.nobs

AverageShiftedHistograms.ash! — Method

ash!(o::Ash; kw...)
ash!(o::Ash, newdata; kw...)
ash!(o::Ash, newdata, weight::AbstractWeights; kw...)

Update an Ash estimate with new data, new weight, smoothing parameter (keyword m), or kernel (keyword kernel):

AverageShiftedHistograms.ash — Method

Univariate Ash

ash(x; kw...)

Fit an average shifted histogram to data x.

ash(x, weight::AbstractWeights; kw...)

Fit a weighted average shifted histogram to data x.

Keyword options are:

rng : values over which the density will be estimated

m : Number of adjacent histograms to smooth over

kernel : kernel used to smooth the estimate

Bivariate Ash

ash(x, y; kw...)

Fit a bivariate averaged shifted histogram to data vectors x and y. Keyword options are:

rngx : x values where density will be estimated

rngy : y values where density will be estimated

mx : smoothing parameter in x direction

my : smoothing parameter in y direction

kernelx : kernel in x direction

kernely : kernel in y direction

Mutating an Ash object

Ash objectes can be updated with new data, new weights(only for univariates), smoothing parameter(s), or kernel(s). They cannot, however, change the ranges over which the density is estimated. It is therefore suggested to err on the side of caution when choosing data endpoints.

# univariate
ash!(obj; kw...)
ash!(obj, newx; kw...)
ash!(obj, newx, weight::AbstractWeights; kw...)

# bivariate
ash!(obj; kw...)
ash!(obj, newx, newy; kw...)

AverageShiftedHistograms.cdf — Method

cdf(o::Ash, x::Real)

Return the estimated cumulative density at the point x.

AverageShiftedHistograms.extendrange — Function

extendrange(x, s = .5, n = 200)

Create a range of length n starting at s standard deviations below minimum(x) and ending at s standard deviations above maximum(x)

AverageShiftedHistograms.histdensity — Method

return the histogram values as a density (integrates to 1)

AverageShiftedHistograms.nout — Method

return the number of observations that fell outside of the histogram range

AverageShiftedHistograms.pdf — Method

pdf(o::Ash, x::Real)

Return the estimated density at the point x.

AverageShiftedHistograms.pdf — Method

pdf(o::Ash2, x::Real, y::Real)

Return the estimated density at the point (x, y).

AverageShiftedHistograms.xy — Method

return the range and density of a univariate ASH

AverageShiftedHistograms.xyz — Method

return ranges and density of biviariate ASH

StatsAPI.nobs — Method

return the number of observations